JSString

JSString

Public API

WebIO.@js_str — Macro.

@js_str(s)

Create a JSString using a string literal and perform interpolations from Julia.

Examples

julia> mystr = "foo"; mydict = Dict("foo" => "bar", "spam" => "eggs");
julia> println(js"const myStr = $mystr; const myObject = $mydict;")
const myStr = "foo"; const myObject = {"spam":"eggs","foo":"bar"};

source

Internal API

Missing docstring.

Missing docstring for WebIO.JSString. Check Documenter's build log for details.

WebIO.tojs — Function.

tojs(x)

Returns a JSString object that constructs the same object as x

source