Render

Render

The WebIO.render function is the primary method of extending WebIO and providing interoperability between WebIO and other libraries. For example, one could define a custom method to render a MyPlot type. See the Extending WebIO documentation for more information.

Internal API

WebIO.register_renderable — Function.

`WebIO.register_renderable(MyType::Type)`

Registers that a WebIO.render method is available for instances of MyType. Allows WebIO to hook into the display machinery of backends such as Atom and IJulia to display the WebIO rendered version of the type as appropriate.

Also defines a Base.show(io::IO, m::MIME"text/html", x::MyType) as Base.show(io, m, WebIO.render(x))

source

WebIO.render — Function.

render(x::MyType)

Generic function that defines how a Julia object is rendered. Should return a Node object.

Examples

struct MyPlot
    s::Scope
end

WebIO.render(p::MyPlot) = WebIO.render(p.s)

source

Private API

WebIO.observable_to_scope — Function.

Wrap an observable in a scope to enable "live updating."

This method also contains distinct code paths for the cases where the observable contains "non-simple" data types (in particular, observables that contain Nodes, Scopes, or Widgets need specially handling).

source